Convert C to C++ code
Replace rw_rewrite_msg with rfc2045::entity::autoconvert, and keep
only rw_rewrite_header().

Make config_defaultdomain() and config_me() more C++-like.

Defensive tactic: before calling idna_to* initialize the return pointer
to NULL, and defensively check the return pointer's value irrespective of
the return code.

+++ courier/courier/libs/readfile.c
@@ -25,7 +25,7 @@ char	*p=0;
 
 	if (fstat(fileno(fp), stat_buf) ||
 		fread((p=(char *)courier_malloc(stat_buf->st_size+1)),
-			1, stat_buf->st_size, fp) != stat_buf->st_size)
+		      1, stat_buf->st_size, fp) != (size_t)stat_buf->st_size)
 	{
 		fclose(fp);
 		clog_msg_errno();
